create responsive module
