+++ web/index.js
@@ -20,7 +20,8 @@ const accordion = new EasyTabAccordion({
     receiver: '[data-accordion-receiver]',
     receiverAttr: 'data-accordion-receiver',
     allowCollapseAll: true,
-    allowExpandAll: true
+    allowExpandAll: true,
+    liveBreakpoint: [1920, 1024]
     //hash: true
 });
